LitBackup
Non-recursive directories fixes

+++ system/filelist-generator.sh
@@ -44,16 +44,6 @@ if [ $STATUS -ne 0 ] ; then
     exit $STATUS
 fi
 
-#Generating a list of files not to recurse
-> $BACKUPROOT/$SERVER-no-recursion-list
-while read line || [[ -n $line ]]; do
-    ssh -i $SSHPRIVATEKEY -p $SSHPORT -o StrictHostKeyChecking=no $SSHUSER@$SERVER "find $line -maxdepth 0 -mindepth 0 -type d" < /dev/null >> $BACKUPROOT/$SERVER-no-recursion-list
-done < $DIRSNORECURSIONLIST
-
-echo "made a list of directories NOT to recurse:"|log
-cat $BACKUPROOT/$SERVER-no-recursion-list | log
-
-
 #Initial values and setup:
 PREVIOUS="JHSdiuagiKUJG45IUYtsok345jhxcgkUH6537648tiyKJHGKu"  #To prevent matching of the first line 
 > $FILESLIST
@@ -64,13 +54,28 @@ chmod 600 $FILESLIST
 sort $FILESLIST.tmp > $FILESLIST.tmp2
 mv -f $FILESLIST.tmp2 $FILESLIST.tmp
 
-#Excluding non recursive dirs from backup:
-grep -Fxv -f $BACKUPROOT/$SERVER-no-recursion-list $FILESLIST.tmp > $FILESLIST.tmp2
-mv -f $FILESLIST.tmp2 $FILESLIST.tmp
 
-#Doing so for rsync - trailing slash copies contents, not the directory itself:
-awk '{print $0"/"}' $BACKUPROOT/$SERVER-no-recursion-list > $BACKUPROOT/$SERVER-no-recursion-list.tmp
-mv -f $BACKUPROOT/$SERVER-no-recursion-list.tmp $BACKUPROOT/$SERVER-no-recursion-list
+if [ -f $DIRSNORECURSIONLIST ] ; then
+	#Generating a list of files not to recurse
+	> $NORECURSEDIRS
+	while read line || [[ -n $line ]]; do
+		ssh -i $SSHPRIVATEKEY -p $SSHPORT -o StrictHostKeyChecking=no
+		$SSHUSER@$SERVER "[ -d $line ] && find $line -maxdepth 0 -mindepth 0 -type d" < /dev/null >> $NORECURSEDIRS ; STATUS=$?
+	done < $DIRSNORECURSIONLIST
+
+	echo "made a list of directories NOT to recurse:"|log
+	cat $NORECURSEDIRS | log
+
+
+
+	#Excluding non recursive dirs from backup:
+	grep -Fxv -f $NORECURSEDIRS $FILESLIST.tmp > $FILESLIST.tmp2
+	mv -f $FILESLIST.tmp2 $FILESLIST.tmp
+
+	#Doing so for rsync - trailing slash copies contents, not the directory itself:
+	awk '{print $0"/"}' $NORECURSEDIRS > $NORECURSEDIRS.tmp
+	mv -f $NORECURSEDIRS.tmp $NORECURSEDIRS
+fi
 
 #Cleaning the file list to only contain parent directories (e.g. /dir and /dir/subdir will leave only /dir - deduplicating to speed up recursive rsync):
 cat $FILESLIST.tmp|while read line
